Update deployed-package-menu by extracting the logic for removing a package into the RemoveDeployedPackageDialog component.
