+++ src/ui/lib/components/DeployedPackageMenu/remove-deployed-package-dialog.svelte
@@ -1,59 +1,27 @@
-<!-- 
-// SPDX-License-Identifier: Apache-2.0
-// SPDX-FileCopyrightText: 2021-Present The Zarf Authors
- -->
 <script lang="ts">
-	import { goto } from '$app/navigation';
-	import { Kind, type DeployedPackage } from '$lib/api-types';
-	import {
-		IconButton,
-		ListItem,
-		ListItemAdornment,
-		Menu,
-		Typography,
-		TextField,
-		type SSX,
-	} from '@ui';
-	import { current_component } from 'svelte/internal';
+	import { Typography, TextField, ListItem, ListItemAdornment } from '@ui';
 	import { Packages } from '$lib/api';
-	import ZarfDialog from './zarf-dialog.svelte';
-	import ButtonDense from './button-dense.svelte';
-	import Spinner from './spinner.svelte';
+	import ZarfDialog from '../zarf-dialog.svelte';
+	import ButtonDense from '../button-dense.svelte';
+	import Spinner from '../spinner.svelte';
+	import type { DeployedPackage } from '$lib/api-types';
 
 	export let pkg: DeployedPackage;
 
-	let anchorRef: HTMLButtonElement;
-	let toggleDialog: () => void;
+	export let toggleDialog: () => void;
 	let inputValue: string;
 	let errorMessage: string;
 	let happyZarf: boolean;
 	let titleText: string;
 	let zarfAlt: string;
 
-	let updateLink = '';
-	let toggled = false;
 	let removing = false;
-
-	const menuSSX: SSX = {
-		$self: {
-			'& .list-item-adornment': { color: 'var(--text-secondary-on-dark)' },
-			'& .divider': { height: '1px', boxShadow: 'inset 0px -1px 0px rgba(255,255,255,0.12)' },
-		},
-	};
 
 	function closeDialog() {
 		inputValue = '';
 		errorMessage = '';
 		removing = false;
 		toggleDialog();
-	}
-
-	function handleClick(event: any) {
-		if (anchorRef && !anchorRef.contains(event.target) && !event.defaultPrevented) {
-			if (anchorRef !== event.target) {
-				toggled = false;
-			}
-		}
 	}
 
 	async function removePkg(): Promise<void> {
@@ -67,13 +35,6 @@
 	}
 
 	$: {
-		if (pkg.data.kind === Kind.ZarfInitConfig) {
-			updateLink = '/packages?init=true';
-		} else {
-			updateLink = '/packages';
-		}
-	}
-	$: {
 		if (!errorMessage) {
 			happyZarf = true;
 			titleText = 'Remove Package from Cluster';
@@ -86,29 +47,6 @@
 	}
 </script>
 
-<svelte:window on:click={handleClick} />
-<IconButton
-	eventComponent={current_component}
-	bind:ref={anchorRef}
-	toggleable
-	bind:toggled
-	on:click={() => (toggled = !toggled)}
-	iconClass="material-symbols-outlined"
-	iconContent="more_vert"
-/>
-<Menu ssx={menuSSX} bind:anchorRef open={toggled} anchorOrigin="bottom-end">
-	<ListItem text="Update Package..." on:click={() => goto(updateLink)}>
-		<ListItemAdornment slot="leading-adornment" class="material-symbols-outlined">
-			cached
-		</ListItemAdornment>
-	</ListItem>
-	<div class="divider" />
-	<ListItem text="Remove..." on:click={toggleDialog}>
-		<ListItemAdornment slot="leading-adornment" class="material-symbols-outlined">
-			delete
-		</ListItemAdornment>
-	</ListItem>
-</Menu>
 <ZarfDialog clickAway={!removing} bind:toggleDialog {happyZarf} {titleText} {zarfAlt}>
 	{#if !errorMessage}
 		{#if removing}
